+++ lib/src/models/package_info_model.dart
@@ -3,11 +3,13 @@ class PackageInfoData {
   String? versionName;
   String? packageName;
   String? versionCode;
+  int? androidSDKVersion;
 
   PackageInfoData({
     this.appName,
     this.versionName,
     this.packageName,
     this.versionCode,
+    this.androidSDKVersion,
   });
 }