+++ lib/src/widgets/placeholder_widget.dart
@@ -3,22 +3,6 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:nb_utils/nb_utils.dart';
 
-/// Colors for PlaceHolderWidget
-List<Color> _placeholderColors = [
-  mistyRose,
-  whiteSmoke,
-  linen,
-  Color(0xFFb7efc5),
-  Color(0xFFf0efeb),
-  Color(0xFFcddafd),
-  Color(0xFFe3d5ca),
-  Color(0xFFffccd5),
-  Color(0xFFccdbdc),
-  Color(0xFFfbf8cc),
-  Color(0xFFcdeac0),
-  Color(0xFFefe5dc),
-];
-
 /// Set nicely colored PlaceHolder while image is loading
 class PlaceHolderWidget extends StatelessWidget {
   final double? height;
@@ -58,9 +42,8 @@ class PlaceHolderWidget extends StatelessWidget {
       duration: animationDuration ?? 1.seconds,
       width: width,
       decoration: BoxDecoration(
-        color: color ??
-            _placeholderColors[
-                Random.secure().nextInt(_placeholderColors.length)],
+        color:
+            color ?? lightColors[Random.secure().nextInt(lightColors.length)],
         shape: shape ?? BoxShape.rectangle,
         borderRadius: borderRadius,
         border: border,