+++ lib/src/widgets/size_listener.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
+import 'package:nb_utils/nb_utils.dart';
 
 /// [SizeListener] Listen to your child widget's size
 class SizeListener extends StatefulWidget {
@@ -25,7 +25,7 @@ class _SizeListenerState extends State<SizeListener> {
 
   @override
   void initState() {
-    SchedulerBinding.instance.addPostFrameCallback(postFrameCallback);
+    afterBuildCreated(() => postFrameCallback);
     super.initState();
   }
 
@@ -43,8 +43,14 @@ class _SizeListenerState extends State<SizeListener> {
   }
 
   @override
+  void didUpdateWidget(covariant SizeListener oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    afterBuildCreated(() => postFrameCallback);
+  }
+
+  @override
   Widget build(BuildContext context) {
-    SchedulerBinding.instance.addPostFrameCallback(postFrameCallback);
+    //SchedulerBinding.instance.addPostFrameCallback(postFrameCallback);
     return Container(
       key: widgetKey,
       child: widget.child,